# contracts/openapi-admin.yaml
openapi: 3.1.0
info:
  title: Myunivokai Admin API
  version: 1.0.0
  description: >
    The internal admin API, staff only. Served at /api/admin/* by the same
    api-gateway process as the public API, but as a separately configured
    route group with its own CORS policy, its own rate limits and
    default-deny authorization. This document is deliberately never merged
    into contracts/openapi.yaml so the public API spec never advertises
    admin routes. Later Sprint 4 phases add the record-browsing and
    analytics-backed routes here.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: "{protocol}://{host}"
    description: Configurable gateway endpoint
    variables:
      protocol:
        default: http
        enum: [http, https]
      host:
        default: localhost:41800
security: []
paths:
...
  /api/admin/worlds:
    get:
      summary: Browse projected worlds
      description: >-
        Cursor paginated. Pagination is mandatory and keyset-based, never
        OFFSET, so every page costs the same and the response stays inside the
        gateway's request/reply deadline as the table grows.
      operationId: adminAnalyticsListWorlds
      security:
        - accessTokenCookie: []
      parameters:
        - { $ref: "#/components/parameters/Family" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/PageSize" }
        - name: archetype
          in: query
          required: false
          schema: { type: string }
        - name: worldStyle
          in: query
          required: false
          schema: { type: string, enum: [cosmic-galaxy, nebula, crystal, aurora, cyber-orbit] }
        - name: mood
          in: query
          required: false
          schema: { type: string }
        - name: published
          in: query
          required: false
          description: Omit for every world; true or false to filter.
          schema: { type: boolean }
        - name: rareFeature
          in: query
          required: false
          description: >-
            Selects the worlds whose stored rarity draw came in under this
            feature's CURRENT probability — the same predicate the overview's
            rarity panel counts with, so a count and the list behind it can
            never disagree. An unrecognised key matches nothing rather than
            being ignored.
          schema:
            type: string
            enum:
              [
                meteor-shower,
                binary-sun,
                black-hole,
                forest-special-bird,
                forest-special-animal,
              ]
      responses:
        "200":
          description: One page of worlds
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorldPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/admin/worlds/{worldID}:
    get:
      summary: Read one projected world and its job history
      description: >-
        Returns the world together with every job that touched it, newest
        first. The two travel in one response because they are read together:
        fetching them separately would let the page show a world beside a job
        list taken from an older read. A 404 covers both "never projected" and
        "not a world id at all" — a projection is eventually consistent, so a
        world that exists upstream but has not arrived here yet is legitimately
        absent.
      operationId: adminAnalyticsGetWorld
      security:
        - accessTokenCookie: []
      parameters:
        - name: worldID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The world and its jobs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorldDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /api/admin/telemetry/routes:
    get:
      summary: Per-route request count, error rate and latency
      description: >-
        One row per route TEMPLATE and method — `/api/universe/worlds/{worldID}`,
        never a path carrying a world id. That rule is what keeps this table
        bounded by the gateway's route count rather than by traffic, which is
        also why it needs no pagination.
      operationId: adminTelemetryRoutes
      security:
        - accessTokenCookie: []
      parameters:
        - { $ref: "#/components/parameters/TelemetryHours" }
      responses:
        "200":
          description: Every route observed in the window, busiest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TelemetryRouteList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/Unavailable" }
components:
  parameters:
    Family:
      name: family
      in: query
      required: false
      description: Omit for every family.
      schema: { type: string, enum: [universe, nature] }
    TelemetryHours:
      name: hours
      in: query
      required: false
      description: >-
        Window in hours, clamped to 1..168 by telemetry-service rather than
        rejected — the maximum matches MYUNIVOKAI_EVENTS's own 7-day retention,
        so asking for more is a question the stream could never have answered.
        Hours rather than days because a bucket here is one minute wide and the
        questions are operational, not historical.
      schema: { type: integer, default: 24, minimum: 1, maximum: 168 }
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque keyset cursor returned as nextCursor by the previous page.
      schema: { type: string }
    PageSize:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: myunivokai_admin_access
      description: >-
        httpOnly, Secure, SameSite=Lax cookie holding the short-lived Ed25519
        access token. The gateway verifies its signature locally and checks
        its tokenVersion against the Redis revocation cache.
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: myunivokai_admin_refresh
      description: >-
        httpOnly, Secure, SameSite=Lax cookie holding the rotating opaque
        refresh token, scoped to the refresh path only.
  schemas:
...
    TraitScores:
      type: object
      additionalProperties: false
      required: [creativity, discipline, curiosity, energy, focus]
      properties:
        creativity: { type: integer, minimum: 0, maximum: 100 }
        discipline: { type: integer, minimum: 0, maximum: 100 }
        curiosity: { type: integer, minimum: 0, maximum: 100 }
        energy: { type: integer, minimum: 0, maximum: 100 }
        focus: { type: integer, minimum: 0, maximum: 100 }
...
    TelemetrySink:
      type: object
      description: >-
        On every telemetry response, not only the ones that fail to answer, so
        the screen reads one field to decide whether to draw charts or a link
        instead of inferring intent from a missing array.
      required: [sink, chartsAvailable]
      properties:
        sink: { type: string, enum: [postgres, otlp] }
        chartsAvailable:
          type: boolean
          description: False when the sink stores nothing locally and cannot answer a range query.
        dashboardUrl:
          type: string
          description: >-
            Where to look instead when chartsAvailable is false. Absent means
            nobody configured one, which the screen states plainly rather than
            rendering a dead link.
...
    TelemetryRouteList:
      allOf:
        - { $ref: "#/components/schemas/TelemetrySink" }
        - type: object
          required: [hours, generatedAt, percentileIsInterpolated, routes]
          properties:
            hours: { type: integer }
            generatedAt: { type: string, format: date-time }
            percentileIsInterpolated: { type: boolean }
            routes:
              type: array
              items:
                type: object
                additionalProperties: false
                required:
                  [
                    routePattern,
                    method,
                    requestCount,
                    errorCount,
                    errorRatePercent,
                    averageDurationMs,
                    p95DurationMs,
                    slowestDurationMs,
                  ]
                properties:
                  routePattern:
                    type: string
                    description: A chi route TEMPLATE, never a raw path.
                  method: { type: string }
                  requestCount: { type: integer }
                  errorCount: { type: integer }
                  errorRatePercent: { type: number }
                  averageDurationMs: { type: integer }
                  p95DurationMs: { type: integer }
                  slowestDurationMs: { type: integer }
...
    World:
      type: object
      additionalProperties: false
      description: >-
        The projected world. Every field here is on the analytics data-boundary
        allow list; the submitted form, the generated profile, the world quote
        and the scene config never cross into this service and can never appear
        on this route.
      required:
        [
          worldId,
          family,
          nickname,
          archetype,
          sceneName,
          mood,
          worldStyle,
          favoriteColors,
          traitScores,
          variantCount,
          selectedVariantNo,
          isPublished,
          revision,
          sourceJobId,
          worldCreatedAt,
          projectedAt,
        ]
      properties:
        worldId: { type: string, format: uuid }
        family: { type: string, enum: [universe, nature] }
        nickname: { type: string }
        role: { type: string }
        archetype: { type: string }
        sceneName: { type: string }
        mood: { type: string }
        worldStyle: { type: string }
        favoriteColors: { type: array, items: { type: string } }
        traitScores: { $ref: "#/components/schemas/TraitScores" }
        variantCount: { type: integer }
        selectedVariantNo: { type: integer }
        isPublished: { type: boolean }
        publishedAt: { type: string, format: date-time }
        revision: { type: integer }
        sourceJobId: { type: string }
        worldCreatedAt: { type: string, format: date-time }
        projectedAt: { type: string, format: date-time }
    WorldPage:
      type: object
      additionalProperties: false
      required: [worlds, totalCount, pageSize]
      properties:
        worlds: { type: array, items: { $ref: "#/components/schemas/World" } }
        nextCursor:
          type: string
          description: Absent on the last page.
        totalCount: { type: integer }
        pageSize: { type: integer }
    WorldDetail:
      type: object
      additionalProperties: false
      required: [world, jobs]
      properties:
        world:
          allOf:
            - $ref: "#/components/schemas/World"
            - type: object
              description: >-
                The list row plus the two identifiers a table has no room for.
                Both are already on the allow list — World carries sourceJobId
                and Job carries profileId and dnaVersionId — so the detail view
                exposes nothing new, it only shows them together.
              properties:
                profileId: { type: string, format: uuid }
                dnaVersionId: { type: string, format: uuid }
        jobs:
          type: array
          description: >-
            Newest first, capped at 50. Joined on world_id rather than followed
            from sourceJobId, because a world accumulates jobs after creation
            and the source job is only the first of them.
          items: { $ref: "#/components/schemas/Job" }
    Job:
      type: object
      additionalProperties: false
      required: [jobId, status, createdAt]
      properties:
        jobId: { type: string }
        family: { type: string, enum: [universe, nature] }
        status: { type: string, enum: [queued, processing, completed, failed] }
        errorCode: { type: string }
        errorMessage: { type: string }
        worldId: { type: string }
        profileId: { type: string }
        dnaVersionId: { type: string }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        durationMs:
          type: integer
          description: Absent while the job is still in flight.
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: array, items: { type: object } }
        requestId: { type: string }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/ErrorBody" }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Missing, invalid, expired or revoked session
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Authenticated, but the account holds none of the required permissions
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: >-
        No such record in the read model. On analytics routes this is not
        necessarily a wrong id: the projection is eventually consistent, so a
        record created moments ago may not have arrived yet.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unavailable:
      description: >-
        auth-service, analytics-service or a required edge dependency could not
        be reached. SERVICE_WAKING means the service was asleep, the gateway has
        started it, and the request never arrived — retry after Retry-After.
        This is the common case on the admin surface, which is used rarely
        enough that its services are usually idle. SERVICE_UNAVAILABLE means a
        real fault that retrying will not fix.
      headers:
        Retry-After:
          description: Seconds to wait before retrying. Sent with SERVICE_WAKING only.
          required: false
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }